pyproject.toml
[project]
name = "smalter-autodoc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"

dependencies = [
    # API
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    # Computer Vision & OCR
    "opencv-python>=4.9.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.2.0",
    "pdf2image>=1.17.0",
    # PDF Handling
    "PyPDF2>=3.0.0",
    "pdfplumber>=0.11.0",
    # Utils
    "python-dotenv>=1.0.0",
    "loguru>=0.7.2",
    "python-magic-bin>=0.4.14",
]


[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "black>=23.12.0",
    "ruff>=0.1.9",
    "mypy>=1.8.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
